requirements.txt
crawl4ai>=0.4.2
python-dotenv>=0.19.0
google-generativeai>=0.3.0
langchain>=0.1.0
chromadb>=0.4.18
tiktoken>=0.5.2
beautifulsoup4>=4.12.2
requests>=2.31.0 